+++ src/index.ts
@@ -1,27 +1,26 @@
-import { default as Application, invokeMiddlewares, Middleware, middlewareCall } from './application';
+export { conditionalCheck } from './conditional';
+
+export { default as Request } from './request';
+export { default as Response } from './response';
+export { default as MemoryRequest } from './memory-request';
+export { default as MemoryResponse } from './memory-response';
+
+export{
+  default as Application,
+  invokeMiddlewares,
+  Middleware,
+  middlewareCall
+} from './application';
+
 export {
   Context,
-  // For backwards compatibility
-  Context as BaseContext,
   WsContext
 } from './context';
-import Headers from './headers';
-import MemoryRequest from './memory-request';
-import MemoryResponse from './memory-response';
-import Request from './request';
-import Response from './response';
-import { conditionalCheck } from './conditional';
 
-export default Application;
 export {
-  Application,
-  conditionalCheck,
   Headers,
-  invokeMiddlewares,
-  middlewareCall,
-  Middleware,
-  Request,
-  Response,
-  MemoryRequest,
-  MemoryResponse,
-};
+  HeadersInterface,
+  HeadersObject
+} from './headers';
+
+export * as headerHelpers from './header-helpers';